+++ src/app/locales/en/components/filesystem-browser/table-row-columns.js
@@ -1,9 +1,9 @@
 export default {
   copy: 'copying',
   move: 'moving',
-  noSizeInfo: 'Directory size statistics are disabled for this space',
+  noSizeInfo: 'Directory statistics are disabled for this space',
   initializingSizeInfo: 'Requested directory statistics are not ready yet – calculation is in progress.',
-  initializingStatsInfo: 'Directory size statistics are being initialized',
+  initializingStatsInfo: 'Directory statistics are being initialized',
   enabledStatsInfo: 'Show more size statistics',
-  disabledStatsInfo: 'Directory size statistics are disabled',
+  disabledStatsInfo: 'Directory statistics are disabled',
 };